refactor: restructure the repo to match project specs and fix all imports.
