+++ utils/file_tracker.py
@@ -20,9 +20,9 @@ from __future__ import annotations
 import hashlib
 import os
 
-from pipeline.utils.db import get_cursor, get_dict_cursor
-from pipeline.utils.retry import db_retry
-from pipeline.logging.logger import get_logger_name
+from warehouse.connection import get_cursor, get_dict_cursor
+from utils.retry import db_retry
+from utils.logger import get_logger_name
 
 logger = get_logger_name(__name__)
